api updated and revoke function added to main login flow as well
